requirements.txt
APScheduler==3.9.1
bcrypt==3.2.2
bidict==0.22.0
blinker==1.4
cffi==1.15.0
click==8.1.3
colorama==0.4.4
DateTime==4.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.2.1
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy==1.22.4
opencv-python==4.5.5.64
Pillow==9.1.0
pycparser==2.21
python-engineio==4.3.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzbar==0.1.9
qrcode==7.3.1
schedule==1.1.0
six==1.16.0
SQLAlchemy==1.4.36
tzdata==2022.1
tzlocal==4.2
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zope.interface==5.4.0
